+++ fe2o3-amqp-types/src/messaging/mod.rs
@@ -1,7 +1,6 @@
 //! Types defined in AMQP 1.0 specification Part 3: Messaging
 
 use serde::{Deserialize, Serialize};
-use serde_amqp::described::Described;
 use serde_amqp::extensions::TransparentVec;
 use serde_amqp::primitives::{Array, OrderedMap};
 use serde_amqp::{primitives::Symbol, value::Value};
@@ -43,7 +42,10 @@ pub use lifetime_policy::*;
 pub type Batch<T> = TransparentVec<T>;
 
 /// 3.5.8 Filter Set
+///
+/// ```xml
 /// <type name="filter-set" class="restricted" source="map"/>
+/// ```
 ///
 /// A set of named filters. Every key in the map MUST be of type symbol,
 /// every value MUST be either null or of a described type which provides
@@ -58,7 +60,14 @@ pub type Batch<T> = TransparentVec<T>;
 /// A registry of commonly defined filter types and their capabilities is
 /// maintained \[AMQPFILTERS\].
 ///
-pub type FilterSet = OrderedMap<Symbol, Option<Described<Value>>>;
+/// Although the value of each entry must be either null or a described type,
+/// many implementations in other programming languages still supports the legacy
+/// format of a filter-set where the value could be any AMQP type. This legacy
+/// format is deprecated and SHOULD NOT be used in new implementations. However,
+/// for compatibility with existing implementations, the following type alias
+/// is provided since "0.7.0" to allow the legacy format to be used.
+pub type FilterSet = OrderedMap<Symbol, Value>;
+// pub type FilterSet = OrderedMap<Symbol, Option<Described<Value>>>;
 
 use crate::definitions::Fields;
 
@@ -134,7 +143,6 @@ impl From<SupportedDistModes> for Fields {
         let mut map = Fields::new();
         let values = modes
             .0
-             .0
             .into_iter()
             .map(Symbol::from)
             .map(Value::Symbol)
@@ -148,7 +156,6 @@ impl From<SupportedDistModes> for Value {
     fn from(modes: SupportedDistModes) -> Self {
         let values: Array<Value> = modes
             .0
-             .0
             .into_iter()
             .map(Symbol::from)
             .map(Value::Symbol)